+++ CHANGELOG.md
@@ -1,4 +1,36 @@
-## 4.0.11 (2020-08-03)
+
+## 4.0.14 (2020-08-13)
+
+#### Enhancements
+* `matchbox`
+  * [#586](https://github.com/SparkPost/matchbox/pull/586) UX-297 Support color in Page primary action ([@jonambas](https://github.com/jonambas))
+
+#### Bug Fixes
+* `matchbox`
+  * [#579](https://github.com/SparkPost/matchbox/pull/579) UX-291 Fix global anchor tag hover styles ([@jonambas](https://github.com/jonambas))
+
+#### Committers: 2
+- Jon Ambas ([@jonambas](https://github.com/jonambas))
+- Logan Sparlin ([@logansparlin](https://github.com/logansparlin))
+
+
+
+
+
+## 4.0.13 (2020-08-04)
+
+#### Features
+
+- `matchbox-icons`, `matchbox-media`
+  - [#568](https://github.com/SparkPost/matchbox/pull/568) UX-272 Create @sparkpost/matchbox-media
+    ([@logansparlin](https://github.com/logansparlin))
+
+#### Committers: 2
+
+- Jon Ambas ([@jonambas](https://github.com/jonambas))
+- Logan Sparlin ([@logansparlin](https://github.com/logansparlin))
+
+## 4.0.13 (2020-08-03)
 
 #### Features